+++ packages/flutter/lib/gestures.dart
@@ -12,10 +12,12 @@ export 'src/gestures/binding.dart';
 export 'src/gestures/constants.dart';
 export 'src/gestures/converter.dart';
 export 'src/gestures/drag.dart';
+export 'src/gestures/drag_details.dart';
 export 'src/gestures/events.dart';
 export 'src/gestures/hit_test.dart';
 export 'src/gestures/long_press.dart';
 export 'src/gestures/lsq_solver.dart';
+export 'src/gestures/monodrag.dart';
 export 'src/gestures/multidrag.dart';
 export 'src/gestures/multitap.dart';
 export 'src/gestures/pointer_router.dart';
